所有NULL更换为nullptr，GCC要求更严格。 Sys.h头文件不再包含string.h，禁止SmartOS除了Type和String以外的地方使用内存管理函数

+++ Net/Ethernet.h
@@ -177,7 +177,7 @@ typedef struct _TCP_HEADER
 
 	void Init(bool recursion = false)
 	{
-		Length = sizeof(this[0]);
+		Length = (byte)sizeof(this[0]);
 		reserved_1 = 0;
 		reserved_2 = 0;
 		//WindowSize = _REV16(8192);
